# Line legend performance